Initial version of several files
I used some of these back when I worked for Design Data, but I've tried
to clean out anything that was proprietary to or owned by the company.
I write tools like this for both personal and work use, so I'm pretty
sure that these are OK to post on GitHub.